2.babymnist/requirements.txt
Flask==1.0.2
Keras==2.0.2
numpy==1.11.3
scipy==0.19.0
tensorflow==1.0.1
Pillow==4.0.0
h5py==2.7.0
line-bot-sdk